FEAT - Menu logic / Counter object / Global refact

+++ headers/Physics/SpriteDirectional3DBody.h
@@ -12,8 +12,8 @@
 class SpriteDirectional3DBody : public SpriteDirectional3D, public Body {
 public:
     btRigidBody* makeRigidBody(float, std::vector<Object3D*> &, Camera3D *, btDiscreteDynamicsWorld*, bool applyCameraImpulse, int forceImpulse);
-    btRigidBody* makeProjectileRigidBody(float, std::vector<Object3D*> &, Camera3D *, btDiscreteDynamicsWorld*, bool applyCameraImpulse, int forceImpulse);
-    btRigidBody* makeProjectileRigidBodyToPlayer(float mass, std::vector<Object3D*> &gameObjects, Vertex3D dir, btDiscreteDynamicsWorld*, int forceImpulse);
+    btRigidBody* makeProjectileRigidBody(float, std::vector<Object3D*> &, Camera3D *, btDiscreteDynamicsWorld*, bool applyCameraImpulse, float forceImpulse);
+    btRigidBody* makeProjectileRigidBodyToPlayer(float mass, std::vector<Object3D*> &gameObjects, Vertex3D dir, btDiscreteDynamicsWorld*, float forceImpulse);
 
     void integrate();
